+++ packages/legacy-scripting-runner/test/example-app/index.js
@@ -642,6 +642,12 @@ const legacyScriptingSource = `
         return bundle.request;
       },
 
+      movie_pre_write_request_data_empty_string: function(bundle) {
+        bundle.request.url = '${AUTH_JSON_SERVER_URL}/echo';
+        bundle.request.data = '';
+        return bundle.request;
+      },
+
       movie_write_default_headers: function(bundle) {
         bundle.request.url = '${HTTPBIN_URL}/post';
         bundle.request.data = z.JSON.stringify({